aaaaah, worst bug ever:
was stupidly doing pretty much a rmtree() on the Hellanzb.QUEUE_DIR
during the clear call. hellanzb has been advertised as able to work
on a public-type QUEUE_DIR (such as your web browser's download
directory, this was even suggested in our OS X howto)

So the clear call could easily have deleted your possibly important
files in that directory, as well as NZBs

I audited the code base to ensure nothing else this stupid was going
on (sorry yak! =[ )